backend/requirements.txt
# Core
django==5.2.11
djangorestframework==3.16.1
mongoengine==0.29.1
django-cors-headers==4.9.0
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
django-filter==25.2

# Async / Realtime
celery==5.6.2
redis==7.2.1
channels==4.3.2
daphne==4.2.1
channels-redis==4.3.0

# Integrations
google-api-python-client
google-auth-oauthlib
stripe==14.4.0
requests==2.31.0
pyotp==2.9.0

# AI/ML
anthropic==0.84.0
scikit-learn==1.7.2
# prophet  # install separately: pip install prophet

# PDF
weasyprint

# Utils
pillow==10.2.0
python-dotenv==1.0.0
